BLD/ENH: Put the team itself inside a submodule.
Various groups have had problems with executing the tests in a standard
way, because it was not possible to say

    $ nosetests test/

from the root of the repository (that way the repository was not seen by
python as a module and therefore was not imported into the tests). The
common workaround to this, doing a `cd ..` before testing, proved highly
impractical and of course cd-ing out of a repository is not an option
when people want to do automatical or at least scripted testing.